+++ packages/types/src/elements/multiSelect/index.ts
@@ -4,11 +4,14 @@ import {
   LezzformElementCategory,
 } from "../..";
 import { ArrayRule } from "../../rules";
+import { ElementAdornmentAttributesType } from "../../shared";
 import { GeneralStyleMetadata } from "../../styles";
 import { DropdownItemApi, DropdownItemStatic } from "../dropdown";
 
 export type MultiSelectType = GenericLezzformElement<
-  GenericFieldAttributes & MultiSelectAttributesType,
+  GenericFieldAttributes &
+    MultiSelectAttributesType &
+    ElementAdornmentAttributesType,
   MultiSelectStyleType
 > & {
   type: "MultiSelect";
@@ -19,7 +22,21 @@ export type MultiSelectType = GenericLezzformElement<
 export type MultiSelectAttributesType =
   | MultiSelectItemApi
   | MultiSelectItemStatic;
-export type MultiSelectStyleType = GeneralStyleMetadata;
+export type MultiSelectStyleType =
+  GeneralStyleMetadata<MultiSelectStyleMetadataValuesType>;
+
+export type MultiSelectStyleMetadataValuesType = {
+  input: Record<string, unknown>;
+  prefixAdornment: {
+    icon?: Record<string, unknown>;
+    text?: Record<string, unknown>;
+  };
+  suffixAdornment: {
+    icon?: Record<string, unknown>;
+    text?: Record<string, unknown>;
+  };
+  content: Record<string, unknown>;
+};
 
 export type MultiSelectRuleType = ArrayRule;